Fix OSX build
