+++ deps/exokit-bindings/videocontext/src/mac/VideoMode_mac.mm
@@ -1,3 +1,4 @@
+#include <VideoCommon.h>
 #include <VideoMode.h>
 #include <AVFoundation/AVFoundation.h>
